Add files via upload
The Attendance Management System is a web-based application designed to simplify and automate the attendance tracking process in educational institutions. This system is structured to enable operators, such as administrative staff, to efficiently manage student attendance, view attendance reports. The system has been developed using a range of modern programming languages and tools, ensuring scalability and ease of use.